정리

+++ src/Controller/API/V1/DeleteResponseController.php
@@ -5,6 +5,7 @@ namespace Lightuna\Controller\API\V1;
 use Lightuna\Controller\AbstractApiController;
 use Lightuna\Controller\AbstractController;
 use Lightuna\Core\Context;
+use Lightuna\Exception\InvalidUserInputException;
 use Lightuna\Http\HttpRequest;
 use Lightuna\Http\HttpResponse;
 use Lightuna\Service\ThreadServiceInterface;
@@ -28,6 +29,16 @@ class DeleteResponseController extends AbstractApiController
     {
         try {
             $logger = $this->context->getLogger();
+
+            $response = $this->threadService->getResponseById($this->input->id);
+            if ($response->getSequence() === 0) {
+                throw new InvalidUserInputException('invalid response sequence');
+            }
+            $thread = $this->threadService->getThreadById($response->getThreadId());
+            if ($thread->getPassword() !== $this->input->password) {
+                throw new InvalidUserInputException('wrong password');
+            }
+
             $this->threadService->deleteResponseById($this->input->id);
             $logger->info("response({$this->input->id}) has been deleted");
             $httpResponse->setBody(json_encode(
